First stub monster implementation. As well as some changes to characters, and addition of the Magic class. Also, broke VC proj into filters (still in progress)

+++ sr2/src/states/SayState.cpp
@@ -59,7 +59,7 @@ void StoneRing::SayState::draw(const CL_Rect &screenRect,CL_GraphicContext * pGC
     pGC->fill_rect( mSpeakerRect, mSpeakerBGColor );
     pGC->fill_rect( mTextRect, mTextBGColor ) ;
 
-    mpSayOverlay->draw(mX,mY, pGC);
+    mpSayOverlay->draw(static_cast<float>(mX),static_cast<float>(mY), pGC);
             
 
     if(miText != mText.end())
